owimageviewer: Move Preview into separate module
